created FormWizard component
